implemented functional chat history loading and saving functionality
the structure of the data in the jsonbin file is provided at sunnymentor-accounts-test.json for your reference. adhere to this structure when reading and writing data from and to the jsonbin file.

chat history loading functionality that occurs when the user opens the chat page:

load the account's entire chat history locally so that if the user loses internet connection they can still locally access previous chats.

currently, the chat history sidebar in the chat page has placeholders "Previous Chat 1" and "Previous Chat 2". rework the chat history sidebar in the frontend to properly display a scrollable list of all the chats in the order that they are listed in the jsonbin file, using the value of the "title" field as the chat's display text.

display an edit icon button on the right of each chat in the chat history sidebar. this edit icon should be invisible normally, and when the user hovers over the chat, it smoothly fades in, and when the user hovers off, it smoothly fades out. when this edit button is clicked, turn the title of the chat in the chat history sidebar into an editable text, allow the user to edit, and then when the user presses enter or clicks somewhere other than the chat title, update the chat title in both the frontend and the jsonbin file and make the chat title no longer editable. when the user clicks the edit button, don't open the chat onscreen, just allow the user to edit the chat name.

create a variable that stores the id number of the chat that the user is currently viewing. if the value of this variable is 0, that means the user is on a new chat and hasn't sent any messages yet. this is a session-specific variable that gets reset back to 0 after the user logs out.

when the user clicks on a chat in the chat history sidebar, set that variable to the id number of the chat that was clicked on. the id number should be equal to the position of the chat's corresponding object in the chat history array in the jsonbin file. now, knowing which chat's messages to display, loop through all message objects from the messages array in the selected chat and display all of them onscreen as sent messages. if the value of a message's "from" field is "user", display it as a user message, and if the value is "chatbot", display it as a chatbot message.

chat history saving functionality that updates the jsonbin file with the new message after every message sent by the user or chatbot:

right after the user or the chatbot sends a message, call a function that appends the new message as an object onto the end of the messages array in the chat array in jsonbin. make sure to append the new message to the right chat based on the current chat id session variable. the new message object should be formatted like this for a message from the user:
{
  "from": "user",
  "content": "contentcontentcontentcontentcontentcontent"
}

and like this for a message from the chatbot:
{
  "from": "chatbot",
  "content": "contentcontentcontentcontentcontentcontent"
}

new chats:

a "new chat" refers to a chat that has been opened but the user hasn't sent any messages yet.

when the user clicks the "new chat" button at the top of the chat history sidebar, set the current chat id variable to 0 and display a blank chat onscreen with only the chatbot's initial message.

upon first opening the chat page, the user should start at a new chat with only the initial message sent by the chatbot.

once the user sends the first message in a new chat, set the chat id variable to 1, then append a new chat object at the beginning of the chat history array in the current account object in the jsonbin file. the new chat object should be formatted like this:

{
  "title": "Chat 1",
  "messages": [
  	{
      "from": "chatbot",
      "content": "[initial greeting message]"
    },
    {
      "from": "user",
      "content": "[user's first message]"
    }
  ]
}

always set the title to the string "Chat 1". always append the newly created chat to the beginning of the chat history array. once the chat object is created, update the message array with the messages currently present in the newly created chat.